[MIG]Models modularization
