+++ l10n_br_hr/models/hr_chronic_disease.py
@@ -1,0 +1,14 @@
+# (c) 2014 Kmee - Luis Felipe Mileo <mileo@kmee.com.br>
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from odoo import api, fields, models
+
+class HrChronicDisease(models.Model):
+    _name = 'hr.chronic.disease'
+
+    name = fields.Char(
+        string='Disease name')
+
+    employee_ids = fields.Many2many(
+        string="Employee",
+        comodel_name='hr.employee')
