+++ l10n_br_hr/models/hr_nationality_code.py
@@ -1,0 +1,24 @@
+# (c) 2014 Kmee - Luis Felipe Mileo <mileo@kmee.com.br>
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from odoo import api, fields, models
+
+
+class HrNationalityCode(models.Model):
+    _name = 'hr.nationality.code'
+
+    name = fields.Char(
+        string='Nationality')
+
+    code = fields.Char(
+        string='Code')
+
+    @api.multi
+    def name_get(self):
+        result = []
+        for record in self:
+            name = record['name']
+            if record['code']:
+                name = record['code'] + ' - ' + name
+            result.append((record['id'], name))
+        return result
